+++ src/migrations/migrate_0_0_10.d.ts
@@ -1,3 +1,0 @@
-declare function migration_0_0_10(): void
-
-export = migration_0_0_10;