+++ src/migrations/migrate_0_0_15.d.ts
@@ -1,3 +1,0 @@
-declare function migration_0_0_15(): void
-
-export = migration_0_0_15;